Tweaking systemd dependencies
Now that the script needs to run at every boot, it is important that the script runs quickly. With the current systemd dependencies, setup-ovs does not run for up to 5+ min after the node has booted during which time the network config is in an incorrect state. This change tweaks the dependencies so the script runs much sooner @ boot.