Swap to balance-alb but continue to use xmit_hash_policy
With balance-xor, we have the issue of broadcast packets looping back to
the bond, and corrupting the bridge's forwarding database, unless we add
some way of filtering out packets that originated from us. While digging
into this, it was discovered that adding that support to xor effectively
turned it into balance-alb, so we gave that a try, and it appears to Just
Work(tm) using balance-alb instead, depsite older documentation claiming
that it can't be used with VMs behind a bridge on top of a bond.

Signed-off-by: Jarod Wilson <jarod@redhat.com>