Linux bridge LB updates
This changes are in the LB.
changed  mode to alb which  makes both sending and receiving frames
load balanced using the change MAC address trick.
Changed tlb_dynamic_lb to 0, which means that load balancing based on
current load is disabled and the load is distributed only
using the hash distribution.