capture-macs, Fix /boot mount handling
the mount command in capture-macs script is flaky on
RHCOS cosa test platform [0], while when using PXE server the
problem does not manifest. The reason is cosa test is faking
things a bit by using a qemu image rather than a real PXE boot.
And in the qemu image, /boot gets mounted very early on,
so this service was racing with that.
In order to allow a more robust capture-macs script, allowing
boot remount should it already been mounted.

[0] https://github.com/coreos/coreos-assembler/issues/2326

Signed-off-by: Ram Lavi <ralavi@redhat.com>