interfaces, change mco manifests to configure basic network
Change machine-config manifests to configure the desired network
configuration.

This commit introduces a new script: init-interfaces, which sets desired
configuration where the default/primary interface is up, and the
secondary is down.
The script is set to run before kubelet.service to avoid potential race with
kubernetes-nmstate configuration on node startup in the future.
This commit also disables the previous setup-ovs.sh script, in order to avoid
conflict when upgrading on existing cluster.

Signed-off-by: Ram Lavi <ralavi@redhat.com>