captured-macs, detach captured-macs.sh fom ignition file
Currently the captured-macs is hard coded to the ignition file.
This makes it annying to test separately from the igntion file.
Detaching the script from ignition.
Renaming the ignition to .tmpl
Refactoring the README to guide how to create the fcc from the
fcc.tmpl file.
Adjusting the ci to do the same.

Signed-off-by: Ram Lavi <ralavi@redhat.com>