tests, handle /boot mount race issues in cosa-ci
In cosa-ci the /boot is artificially mounted in
read-only mode. This behavior is different from how it
is mounted on PXE server, and as a result capture-mac's
independently 'mount /boot' command races with the
regular but read-only boot mount.
Instead, depend on the latter, then remount it read-write
in a mount namespace
This commit mimics a similar fix commit in cosa-ci [0]

[0] https://github.com/coreos/coreos-assembler/pull/2328

Signed-off-by: Ram Lavi <ralavi@redhat.com>