infra: Use kcli to test the repo
With kcli is possible to start a openshift cluster and
apply ignition and also pass kernel cmkdline, this way
is possible to simulate the scenario.

Some tunning is needed to make it work, it has being
included in the change but does not maean that we have to
merge it.

Signed-off-by: ellorent <ellorent@redhat.com>