Detect deactivated profiles based on name/mac
The current way of detecting profiles is not working reliably during
boot. The existing version of get_con_name_by_device is unable to
detect deactivated profiles associated to the device. That leads to the
profile of not being found, we create our own and then conflict with the
old profile getting activated.

This brings back commit ef360ae6d1fc11562c074a1c270766edfe7f5fe1 except
it drops code that introduced explicit MAC settings on profiles we create.

Signed-off-by: Petr Horáček <phoracek@redhat.com>