test/network.go, Fix test for new init-interfaces
Due to change in init-interfaces script to keep the connection up with
specific ipv4 enabled/disabled, the tests needs to change accordingly.
Also, since currently there is need to test upgrade scenario, we can
simpify the tests.

Merged the two current tests
'rhcos.network.init-interfaces-pre-configured' and
'rhcos.network.init-interfaces-not-configured' to
'rhcos.network.init-interfaces-test'.
The new test checks 2 scenarios:
- new setup scenario on fresh node, where there are not connections created.
- setup after node reboot, that makes sure the
interfaces don't change after reboot (when MCO reruns the
init-interfaces script).
The test is also changed to check that the IPs of the interfaces are in
the correct state.

Signed-off-by: Ram Lavi <ralavi@redhat.com>