ci, swtpm: Pin to a supported version ubuntu 20.04
The CI is currently running with the latest ubuntu version (22.04)
The latter does not support the use the swtpm package currently used.
In order to allow the CI to work, temporarilly downgrading the ubuntu
version to the former one 20.04.
This does not affect the actual CI testing as they are run on the RHCOS
VM.

Signed-off-by: Ram Lavi <ralavi@redhat.com>