mco: Set OVS links to MACAddressPolicy none
Due to a bug [0] in networkManager, when set to
MACAddressPolicy=persistent links can get reapplied. This can result
with unexpected behaviors like the node getting a new IP lease from
DHCP.
In order to prevent this, changing the MACAddressPolicy to none on ovs
links.
On newer releases the policy is already set to none [1] so this change
will only matter in old releases.

[0] https://issues.redhat.com/browse/RHEL-40683
[1] https://fedoraproject.org/wiki/Changes/MAC_Address_Policy_none

Signed-off-by: Ram Lavi <ralavi@redhat.com>